Split shell job lifecycle tests
